+++ src/segm8.h
@@ -30,13 +30,13 @@ public:
     void begin();
 
     void clear();
-    void display(int number, uint8_t position, uint8_t width,
+    void display(int16_t number, uint8_t position, uint8_t width,
         uint8_t flags = SEGM8_ALIGN_RIGHT);
-    void display(unsigned int number, uint8_t position, uint8_t width,
+    void display(uint16_t number, uint8_t position, uint8_t width,
         uint8_t flags = SEGM8_ALIGN_RIGHT | SEGM8_RADIX_10);
-    void display(long number, uint8_t position, uint8_t width,
+    void display(int32_t number, uint8_t position, uint8_t width,
         uint8_t flags = SEGM8_ALIGN_RIGHT);
-    void display(unsigned long number, uint8_t position, uint8_t width,
+    void display(uint32_t number, uint8_t position, uint8_t width,
         uint8_t flags = SEGM8_ALIGN_RIGHT | SEGM8_RADIX_10);
     void display(float number, uint8_t position, uint8_t width,
         uint8_t precission = 1, uint8_t flags = SEGM8_ALIGN_LEFT);
